Updated intro (Overview)
I understand that editing the frame text for the content snippet may be out of the bounds of this assessment, but I saw some potential for optimizing these sections and thought I would share some notes. If this is outside the scope of this project, please feel free to skip down to the Content section.

I reframed the opening to clarify the context for this assessment. I recast the first sentence in active voice and reordered it to follow the given-new pattern of ordering information (in this case, the interview process is old information, while introducing this assessment is new). I also grouped like information together by adding the assignment duration to the initial sentence. It's a minor tweak, but it will help set the reader's expectations from the start.