Update bullet 1, list 1
In the previous version of the list, the frame text for the entire list ("Construct your PR to teach the author") didn't work with every item in the list. Typically in a bulleted list, the introductory text is grammatically consistent with every item in the list. In a list, the bullets should also be consisten with one another, in terms of whether they are incomplete phrases or complete sentences (with subjects and verbs). In this case, it seemed like "teaching the author" primarily connected to the idea of atomic commits. So I grouped that information together and added more generic introductory text for the list as a whole. However, I might be misunderstanding the intent here! I'd love to hear your thoughts on this first bullet or the list as a whole.

I also have a question about atomic commits. The first bullet mentions these, but they're not explained. Who is the "author" in this case -- the imagined author of the code snippet or the technical writer completing this assessment? Either way, this concept seems worth explaining for readers. Since it requires more explanation, it might make sense to move the discussion about atomic commits to the "Instructions" section, since it might relate to explaining how to complete the assignment.