Added intro paragraph (Content section)
I suggest adding a brief introduction before the bulleted list defining commands. An introduction (even a few short sentences) helps motivate the reader to keep reading and provides context for the discussion below. Introductions also help readers quickly determine whether the document will answer their questions or meet their needs. Because of this, I suggest adding some introductory text to clarify the purpose and context of this content. For example, what's the use case for needing to distinguish between these Git commands? I added some sentences as an example, but I welcome your input on this content's purpose and primary use case. (For instance, we may not need to begin with an overview of Git.)