added git merge
I added `git merge` as another primary command for synchronizing changes. Merging is also part of git pull discussion below, so it seemed helpful to include it in this overview with the other main commands. However, merging might be outside the scope of this piece -- I welcome your input on this.