Updated opening lines
Currently, the discussion begins with a common misperception, which can be a great hook for motivating readers to dig into the truth behind the perception. In this case, though, I had a little trouble following exactly what the equivalence was between `git push` and `git pull`. For example, are these misunderstood as equivalent because they are equal, but opposite actions (one downloads changes while the other uploads them)? Or, should this read "`git fetch` and `git pull`" because both of those download changes, and could be misunderstood as the same? If this hook is a good fit for the audience and purpose, I suggest clarifying this misperception so that readers can follow along more easily. Another option here would be to remove the misperception and reframe the discussion more positively in terms of what these commands do. This option aligns a little more closely with our style guide, since we might not want to assume what readers don't understand. The simple approach might be to focus on explaining these concepts without the misperception. For now, I've reframed the opening in terms of a "job to be done" by moving one of the later sentences up, but I'd like to know your thoughts on the purpose of this piece and your goals for the opening.