Deleted sentences
I deleted the phrase "make the remote branch resemble my local branch" since it's a little informal for our style, and I think it's more effective to name this task at the start of the paragraph. Knowing the task from the beginning helps the reader follow the discussion of the details.