Added paragraph break and transition
I created a paragraph break here to help signal the transition from when a push works as expected to when it doesn't. Paragraph breaks create smaller, more digestible chunks of content for readers, which also makes the document easier to skim and read. I also added a transition to mark the new idea. This also helps readers skim to find the information they need. I also replaced "this" with "the push" because the referent for "this" might be unclear to readers.