Copyedit for clarity
In the previous version, the clause following the colon was a little hard to follow. Usually when readers see a statement beginning with "if", they expect to see a corresponding "then". In this case, creating a separate sentence and expanding it to include the consequence (it will fail) helps clarify the meaning for readers. What do you think about adding the error message for a failed push here? That could help with SEO/discoverability.