Updated commands
For consistency, I set the commands in code font. I also changed the order of the commands and added a comma to clarify. In this case, the "either/or" construction is a little more complicated since one of the options is actually two commands used together. (So presenting these in a straight list could lead to confusion for learners.) Reversing the order and adding the comma will help differentiate the two approaches. I also replaced "and" with "with" to clarify that fetch and merge need to be used together (in this use case).