Copyedit with question
For clarity, I moved the mentions of branch names into separate sentences, which creates more space for explanation. Can you clarify why it matters that both branches may be called "master"? In the context of this discussion and the `git fetch` command, it's not entirely clear why the reader needs to know that their local branch may have almost the same name as the remote branch. If the branch names are not directly relevant to understanding how git pull works, I might suggest removing the sentence about the local branch being called "master". (It makes sense to keep the remote branch name, since that's the argument for the `git fetch` command example.) However, if these names are relevant, I suggest expanding on this to explain the connection to readers. On a related note, will these naming conventions be true for all readers? From what I understand, Git uses "master", while GitHub uses "main". If these naming conventions are useful to know (for some readers), but not directly relevant for everyone, perhaps we could move this discussion to a note. That way, it's there if readers need it, but more advanced readers could also skip over it if not. Alternatively, we could also link to a relevant resource here. I welcome your input on this!