Updated final sentence
The previous sentence helpfully described if/when to use each of the approaches discussed. However, it was a little unclear what the main conclusion for readers might be. What do you think about taking this conclusion a step further with a clearer recommendation? To test this out, I recast the sentence as a "recommendation" to clarify the key takeaway for readers. However, I might be misunderstanding the intent here (and I might be mispresenting this as HashiCorp's recommendation). In your view, what's the key takeaway for readers? Is using `git fetch` and `git merge` a best practice? Why/why not? Or is your goal to educate readers about when to use each approach? Depending on the purpose/audience for this document, I suggest expanding on this point with additional discussion and use cases. For example, what are the values, risks, and use cases for each approach (`git fetch` and `git merge` VS `git pull`)?